sdm632-common: Remove ineffective BOARD_KERNEL_TAGS_OFFSET variable
Remove various variables related to where kernel and ramdisk are loaded
by the bootloader. There are two reasons for this:

1. moto_sdm632 bootloader ignores all the offset values for kernel, ramdisk
and tags.

2. The variables BOARD_KERNEL_TAGS_OFFSET and BOARD_RAMDISK_OFFSET are
ignored by the Android build system (unless they are copied into the
BOARD_MKBOOTIMG_ARGS variable). This is in contrast to the variables
BOARD_KERNEL_BASE, BOARD_KERNEL_PAGESIZE and BOARD_MKBOOTIMG_ARGS which
are referenced by build/make/core/Makefile

Bug: 117832711
Change-Id: I3404a4e084ed063d51077a3a80229854362c768b